Clear memory area for BSS and enable SSE types of instructions
The BSS section isn't cleared by default after the kernel has been
loaded. This is now done during start of the kernel to ensure that
memory there is initialized with 0. For this we have added two
linker variables that we use to know where the BSS exist in memory.

Also enable SSE and SSE2 instructions. These instructions are used
when you compile the kernel with -O2 and higher optimization level. If
not enabled this will cause a crash.